### 2021-04-22 #3208 @whedon
[PRE REVIEW]: MLJFlux: Deep learning interface to the MLJ toolbox
**Submitting author:** @ayush-1506 (<a href="http://orcid.org/0000-0003-3550-691X">Ayush Shridhar</a>)
**Repository:** <a href="https://github.com/FluxML/MLJFlux.jl" target ="_blank">https://github.com/FluxML/MLJFlux.jl</a>
**Version:** v0.1.10
