### 2021-04-28 a96db02
CompatHelper: bump compat for "ColorTypes" to "0.11"
