+++ Project.toml
@@ -16,7 +16,7 @@ Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
 CategoricalArrays = "^0.8.1,^0.9"
-ColorTypes = "^0.10.3"
+ColorTypes = "^0.10.3, 0.11"
 ComputationalResources = "^0.3.2"
 Flux = "^0.10.4, ^0.11, 0.12"
 LossFunctions = "^0.5, ^0.6"
