### 2021-05-02 #3208 @danielskatz
@melissawm - Can you explain specifically why you flagged this, and what other editors should look for? It seems like enough code/contribution, and it seems like research software, so I'm unsure what the potential issue is. It would be better if the authors provided some examples of research problems where this could be used, but I'm fairly confident this shouldn't be difficult for them to do.
