+++ paper/paper.bib
@@ -42,7 +42,7 @@
 }
 
 @online{MLJFlux.jl,
-  author={Shridhar, Ayush and Bloam, Anthony},
+  author={Shridhar, Ayush and Blaom, Anthony},
   title = {An interface to the deep learning package Flux.jl from the MLJ.jl toolbox },
   url = {https://github.com/FluxML/MLJFlux.jl},
 }
