### 2021-05-15 a6b9c11
bump compat MLJModelInterface = "1.1"
