+++ Project.toml
@@ -15,15 +15,15 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 
 [compat]
-CategoricalArrays = "^0.10"
-ColorTypes = "^0.10.3, 0.11"
-ComputationalResources = "^0.3.2"
-Flux = "^0.10.4, ^0.11, 0.12"
-LossFunctions = "^0.5, ^0.6"
-MLJModelInterface = "^0.4.1, 1"
-ProgressMeter = "^1.1"
-Tables = "^1.0"
-julia = "^1.3"
+CategoricalArrays = "0.10"
+ColorTypes = "0.10.3, 0.11"
+ComputationalResources = "0.3.2"
+Flux = "0.10.4, 0.11, 0.12"
+LossFunctions = "0.5, 0.6"
+MLJModelInterface = "1.1"
+ProgressMeter = "1.1"
+Tables = "1.0"
+julia = "1.3"
 
 [extras]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
