### 2021-05-15 38c4ac7
update list of optimisers; re-instate optimiser transparency
