### 2021-05-15 1d20458
add :optimiser and :builder to deep_properties to address #129 and #124
