### 2021-05-15 39e5984
update examples/mnist
