+++ examples/mnist/mnist.md
@@ -9,7 +9,11 @@ using Pkg
 const DIR = @__DIR__
 Pkg.activate(DIR)
 Pkg.instantiate()
+```
 
+**Julia version** is assumed to be 1.6.*
+
+```@example mnist
 using MLJ
 using Flux
 import MLJFlux
