+++ Project.toml
@@ -23,7 +23,7 @@ LossFunctions = "0.5, 0.6"
 MLJModelInterface = "1.1"
 ProgressMeter = "1.1"
 Tables = "1.0"
-julia = "1.3"
+julia = "1.6"
 
 [extras]
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
