### 2021-05-15 b16db0a
remove julia <1.6 from CI.yml
