+++ .github/workflows/ci.yml
@@ -17,8 +17,6 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '1.3'
-          - '1.4'
           - '1'
           - 'nightly'
         os:
