### 2021-05-19 ff480a1
CompatHelper: bump compat for "LossFunctions" to "0.7"
