+++ Project.toml
@@ -19,7 +19,7 @@ CategoricalArrays = "0.10"
 ColorTypes = "0.10.3, 0.11"
 ComputationalResources = "0.3.2"
 Flux = "0.10.4, 0.11, 0.12"
-LossFunctions = "0.5, 0.6"
+LossFunctions = "0.5, 0.6, 0.7"
 MLJModelInterface = "1.1"
 ProgressMeter = "1.1"
 Tables = "1.0"
