### 2021-05-27 1ef2661
update notebook, readme
