### 2021-05-28 b388d72
further readme tweaks including forgotten Plots package in env
