### 2021-05-28 cdc5650
generated an executed notebook for MNIST
