### 2021-06-03 8f679e6
Make sure the nightly CI has a different workflow name (#149)
