+++ .github/workflows/ci_nightly.yml
@@ -1,4 +1,4 @@
-name: CI
+name: CI (Julia nightly)
 on:
   pull_request:
     branches:
@@ -10,7 +10,7 @@ on:
       - dev
     tags: '*'
 jobs:
-  test:
+  test-julia-nightly:
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
     runs-on: ${{ matrix.os }}
     strategy:
