### 2021-06-08 5ecbe45
correct the doc-string for `fit!(chain, ...)` which is out-of-date
