+++ test/core.jl
@@ -1,5 +1,8 @@
 Random.seed!(123)
 
+rowvec(y) = y
+rowvec(y::Vector) = reshape(y, 1, length(y))
+
 @test MLJFlux.MLJModelInterface.istransparent(Flux.ADAM(0.1))
 
 @testset "nrows" begin
@@ -19,7 +22,8 @@ end
     model = MLJFlux.NeuralNetworkRegressor()
     model.batch_size= 3
     @test MLJFlux.collate(model, X, y) ==
-        ([Xmatrix'[:,1:3], Xmatrix'[:,4:6], Xmatrix'[:,7:9], Xmatrix'[:,10:10]], [y[1:3], y[4:6], y[7:9], y[10:10]])
+        ([Xmatrix'[:,1:3], Xmatrix'[:,4:6], Xmatrix'[:,7:9], Xmatrix'[:,10:10]],
+         rowvec.([y[1:3], y[4:6], y[7:9], y[10:10]]))
 
     # NeuralNetworClassifier:
     y = categorical(['a', 'b', 'a', 'a', 'b', 'a', 'a', 'a', 'b', 'a'])
@@ -27,8 +31,10 @@ end
     model.batch_size = 3
     data = MLJFlux.collate(model, X, y)
 
-    @test data == ([Xmatrix'[:,1:3], Xmatrix'[:,4:6], Xmatrix'[:,7:9], Xmatrix'[:,10:10]],
-        [[1 0 1; 0 1 0], [1 0 1; 0 1 0], [1 1 0; 0 0 1], reshape([1; 0], (2,1))])
+    @test data == ([Xmatrix'[:,1:3], Xmatrix'[:,4:6],
+                    Xmatrix'[:,7:9], Xmatrix'[:,10:10]],
+                   rowvec.([[1 0 1; 0 1 0], [1 0 1; 0 1 0], [1 1 0; 0 0 1],
+                            reshape([1; 0], (2,1))]))
 
     # MultitargetNeuralNetworRegressor:
     ymatrix = rand(10, 2)
@@ -37,12 +43,14 @@ end
     model.batch_size= 3
     @test MLJFlux.collate(model, X, y) ==
         ([Xmatrix'[:,1:3], Xmatrix'[:,4:6], Xmatrix'[:,7:9], Xmatrix'[:,10:10]],
-                [ymatrix'[:,1:3], ymatrix'[:,4:6], ymatrix'[:,7:9], ymatrix'[:,10:10]])
+         rowvec.([ymatrix'[:,1:3], ymatrix'[:,4:6], ymatrix'[:,7:9],
+                  ymatrix'[:,10:10]]))
 
     y = Tables.columntable(y) # try a columnaccess table
     @test MLJFlux.collate(model, X, y) ==
         ([Xmatrix'[:,1:3], Xmatrix'[:,4:6], Xmatrix'[:,7:9], Xmatrix'[:,10:10]],
-            [ymatrix'[:,1:3], ymatrix'[:,4:6], ymatrix'[:,7:9], ymatrix'[:,10:10]])
+         rowvec.([ymatrix'[:,1:3], ymatrix'[:,4:6],
+                  ymatrix'[:,7:9], ymatrix'[:,10:10]]))
 
     # ImageClassifier
     Xmatrix = coerce(rand(6, 6, 1, 10), GrayImage)
@@ -50,7 +58,8 @@ end
     model = MLJFlux.ImageClassifier(batch_size=2)
 
     data = MLJFlux.collate(model, Xmatrix, y)
-    @test  first.(data) == (Float32.(cat(Xmatrix[1], Xmatrix[2], dims=4)), [1 0;0 1])
[…]
