### 2021-06-08 cf032dc
New function tomat to reshape vectors into matrices
