### 2021-06-08 f4a471a
Use the tomat function in predict to adress #142
