### 2021-06-10 7d4589c
Correclty load MultitargetNeuralNetworkRegressor (#158)
The following
```julia
Model = @load MultitargetNeuralNetworkRegressor pkg=MLJFlux verbosity=0
```
> MLJFlux.NeuralNetworkRegressor
would load a basic `NeuralNetworkRegressor`.
