### 2021-06-14 bf6374d
mv type defs and constructors out to src/types.jl
