+++ src/regressor.jl
@@ -1,84 +1,3 @@
-mutable struct NeuralNetworkRegressor{B,O,L} <: MLJModelInterface.Deterministic
-    builder::B
-    optimiser::O    # mutable struct from Flux/src/optimise/optimisers.jl
-    loss::L         # can be called as in `loss(yhat, y)`
-    epochs::Int     # number of epochs
-    batch_size::Int # size of a batch
-    lambda::Float64 # regularization strength
-    alpha::Float64  # regularizaton mix (0 for all l2, 1 for all l1)
-    optimiser_changes_trigger_retraining::Bool
-    acceleration::AbstractResource       # To use GPU
-end
-
-function NeuralNetworkRegressor(; builder::B   = Linear()
-                                , optimiser::O = Flux.Optimise.ADAM()
-                                , loss::L      = Flux.mse
-                                , epochs       = 10
-                                , batch_size   = 1
-                                , lambda       = 0
-                                , alpha        = 0
-                                , optimiser_changes_trigger_retraining=false
-                                , acceleration  = CPU1()
-                                ) where {B,O,L}
-
-    model = NeuralNetworkRegressor{B,O,L}(builder
-                                          , optimiser
-                                          , loss
-                                          , epochs
-                                          , batch_size
-                                          , lambda
-                                          , alpha
-                                          , optimiser_changes_trigger_retraining
-                                          , acceleration)
-
-   message = clean!(model)
-   isempty(message) || @warn message
-
-    return model
-end
-
-mutable struct MultitargetNeuralNetworkRegressor{B,O,L} <: MLJModelInterface.Deterministic
-    builder::B
-    optimiser::O    # mutable struct from Flux/src/optimise/optimisers.jl
-    loss::L         # can be called as in `loss(yhat, y)`
-    epochs::Int          # number of epochs
-    batch_size::Int # size of a batch
-    lambda::Float64 # regularization strength
-    alpha::Float64  # regularizaton mix (0 for all l2, 1 for all l1)
-    optimiser_changes_trigger_retraining::Bool
-    acceleration::AbstractResource
-end
-
-function MultitargetNeuralNetworkRegressor(; builder::B   = Linear()
-                                           , optimiser::O = Flux.Optimise.ADAM()
-                                           , loss::L      = Flux.mse
-                                           , epochs       = 10
-                                           , batch_size   = 1
-                                           , lambda       = 0
-                                           , alpha        = 0
-                                           , optimiser_changes_trigger_retraining=false
-                                           , acceleration = CPU1()
-                                           ) where {B,O,L}
-
-    model = MultitargetNeuralNetworkRegressor{B,O,L}(builder
-                                                     , optimiser
-                                                     , loss
-                                                     , epochs
-                                                     , batch_size
-                                                     , lambda
-                                                     , alpha
-                                                     , optimiser_changes_trigger_retraining
-                                                     , acceleration)
-
-    message = clean!(model)
-    isempty(message) || @warn message
-
-    return model
-end
-
-const Regressor =
-    Union{NeuralNetworkRegressor, MultitargetNeuralNetworkRegressor}
-
 function MLJModelInterface.fit(model::Regressor, verbosity::Int, X, y)
 
     # (assumes  no categorical features)
