+++ src/classifier.jl
@@ -1,46 +1,3 @@
-mutable struct NeuralNetworkClassifier{B,F,O,L} <: MLJModelInterface.Probabilistic
-    builder::B
-    finaliser::F
-    optimiser::O    # mutable struct from Flux/src/optimise/optimisers.jl
-    loss::L         # can be called as in `loss(yhat, y)`
-    epochs::Int          # number of epochs
-    batch_size::Int # size of a batch
-    lambda::Float64 # regularization strength
-    alpha::Float64  # regularizaton mix (0 for all l2, 1 for all l1)
-    optimiser_changes_trigger_retraining::Bool
-    acceleration::AbstractResource       # Train on gpu
-end
-
-function NeuralNetworkClassifier(; builder::B   = Short()
-                                 , finaliser::F = Flux.softmax
-                                 , optimiser::O = Flux.Optimise.ADAM()
-                                 , loss::L      = Flux.crossentropy
-                                 , epochs       = 10
-                                 , batch_size   = 1
-                                 , lambda       = 0
-                                 , alpha        = 0
-                                 , optimiser_changes_trigger_retraining = false
-                                 , acceleration = CPU1()
-                                 ) where {B,F,O,L}
-
-    model = NeuralNetworkClassifier{B,F,O,L}(builder
-                                             , finaliser
-                                             , optimiser
-                                             , loss
-                                             , epochs
-                                             , batch_size
-                                             , lambda
-                                             , alpha
-                                             , optimiser_changes_trigger_retraining
-                                             , acceleration
-                                             )
-
-   message = clean!(model)
-   isempty(message) || @warn message
-
-    return model
-end
-
 function MLJModelInterface.fit(model::NeuralNetworkClassifier,
                                verbosity::Int,
                                X,

[…]
