### 2021-06-15 07921fd
add abstractions to enable single fit/update for all models
oops
