### 2021-06-15 364fb51
change the signature of build and add user rng interface point
