### 2021-06-15 d3c3a9a
change the signature of build and add user rng interface point
