### 2021-06-15 819881a
use single glorot_uniform function in given builder
