### 2021-06-15 53b5790
use single glorot_uniform function in given builder
