### 2021-06-15 f7dd38a
adjust mnist example in readme
