+++ README.md
@@ -351,8 +351,9 @@ following conditions:
 
 - `chain(x)` must make sense:
 
-    - for any `x <: Vector{<:AbstractFloat}` of length `n_in` (for use
-      with one of the first three model types); or
+    - for any `x <: Array{<:AbstractFloat, 2}` of size `(n_in,
+      batch_size)` where `batch_size` is any integer (for use with one
+      of the first three model types); or
 
     - for any `x <: Array{<:Float32, 4}` of size `(W, H, n_channels,
       batch_size)`, where `(W, H) = n_in`, `n_channels` is 1 or 3, and
