### 2021-06-21 d2b51af
add test (and StableRNGs to [extras])
add StableRNGs for testing

add test
