### 2021-06-23 #3375 @krystophny
@ayush-1506 @melissawm sorry for the delay, I have now finally found time to look at the package in detail. Unfortunately, I have the feeling that this is just a "thin API client" in the sense of the JOSS guidelines.

As far as I understand, the package provides an interface/wrappers between `Flux` and `MLJ`. Four types are exported: `NeuralNetworkRegressor`, `MultitargetNeuralNetworkRegressor`, `NeuralNetworkClassifier` and `ImageClassifier`. In total, those are implemented in less than 600 lines of code (`src` directory), including some boilerplate. All that is done are translations between two APIs. I wouldn't count his as "substantial scholarly effort". Also, the documentation is quite minimalistic. I acknowledge that there are a number of users, which indicates the usefulness of the package. Still, I don't believe it is extensive enough to justify a publication in JOSS.
