### 2021-06-23 43c0857
update builder doc-strings
