### 2021-06-24 74c5e8b
fix fault test leading to "scalar indexing issue" on GPU builds
fix faulty test

revert simpler code for getting initial training loss

oops, remove redundant normalization

fix typo
