### 2021-06-28 #3375 @morganericsson
@melissawm Yes, I agree with @krystophny . I've reviewed the code and played around with it a bit to get a feel for how it works. It provides a way to wrap Flux models and use them using the MLJ API. There are some nice benefits to this, but these are (mainly) due to MLJ and other underlying packages; most of the code in this package is just wrapping and translating (by design).

I found the documentation sufficient. I was able to define a simple builder (based on a Keras tutorial) and use it in a simple pipeline (even if that is mainly Flux and MLJ). However, I think the paper does a poor job of presenting the package and motivating why it is needed. I found, e.g., https://discourse.julialang.org/t/pros-and-cons-of-using-flux-directly-over-using-mlj-mljflux/59756 to describe the benefits much better (and, in some sense, the need).
