### 2021-06-29 36d2eda
address ScientificTypes name change
