### 2021-07-03 8e1e881
add penalized loss function and integrate with fit/update
oops

oops
