### 2021-07-03 cec592f
add penalized loss function and integrate with fit/update
oops

oops

correct order of move and collate

fix missing chain move

fix forgotten addition to cache

update doc-string
