### 2021-07-06 #3375 @morganericsson
@melissawm I am done with the form now. Below are comments for the boxes I did not tick.

- I agree with @krystophny that I would not count this as a "substantial scholarly effort" in the JOSS sense, even if I can see how the package is useful.
- I could not find any community guidelines.
- I would like to see an improved _statement of need_ / _state of the field_ (these are quite similar in this case in my opinion). For example, I think the discussion on when to use MLJFlux, MLJ, Flux, or FastAI could be improved (as I mentioned in my previous comment).
- I think the references could be improved. For example, I am not sure why there is a need to self-reference?

While the documentation and examples are sufficient, I think it would be very helpful to improve these. For example, provided examples (`examples/`) and the examples in README.md are quite similar. It would have been nice with a few more complicated examples, especially a more complicated builder. But, as I said, they were sufficient to get me started, and I suppose the MLJ and Flux documentation becomes more relevant when you start to do more advanced stuff.
