### 2021-07-06 #3375 @melissawm
Thank you both @morganericsson and @krystophny . At this point I'm tagging @openjournals/joss-eics to evaluate next steps.
