+++ README.md
@@ -275,6 +275,9 @@ MLJ provides two simple builders out of the box. In all cases weights
 
 See Table 1 above to see how `n_in` and `n_out` relate to the data.
 
+Alternatively, `MLJFlux.@build(neural_net)` automatically creates a builder
+for any valid `neural_net`, e.g. `MLJFlux.@build(Dense(100, 2, tanh))`.
+
 
 ### Model hyperparameters.
 
@@ -378,6 +381,7 @@ following conditions:
 - The object returned by `chain(x)` must be an `AbstractFloat` vector
   of length `n_out`.
 
+See also `MLJFlux.@build` for an automated way to create new builders.
 
 ### Loss functions
 
