### 2021-07-09 ce0c362
n_in, n_out, n_channels
