### 2021-07-10 68d3de5
add integration test for regularization
