### 2021-07-10 49a3e25
minor opimization for case of zero regularization
