### 2021-07-13 67409a5
use sum(abs2, A) instead of sum(x^2 for x in A) in loss penalty
