+++ .github/workflows/ci.yml
@@ -13,6 +13,8 @@ jobs:
   test:
     name: Julia ${{ matrix.version }} - ${{ matrix.os }} - ${{ matrix.arch }} - ${{ github.event_name }}
     runs-on: ${{ matrix.os }}
+    env:
+      PYTHON: ""
     strategy:
       fail-fast: false
       matrix:
