+++ README.md
@@ -168,7 +168,7 @@ Flux "models" used in MLJFLux are `Flux.Chain` objects, we call them
 MLJFlux provides four model types, for use with input features `X` and
 targets `y` of the [scientific
 type](https://alan-turing-institute.github.io/MLJScientificTypes.jl/dev/)
-indicated in the table below. The parameters `n_in` and `n_out`
+indicated in the table below. The parameters `n_in`, `n_out` and `n_channels`
 refer to information passed to the builder, as described under
 [Defining a new builder](defining-a-new-builder) below.
 
@@ -275,15 +275,6 @@ MLJ provides two simple builders out of the box. In all cases weights
 
 See Table 1 above to see how `n_in` and `n_out` relate to the data.
 
-Alternatively, use `MLJFlux.@builder(neural_net)` to automatically create a builder for
-any valid Flux chain expression `neural_net`, where the symbols `n_in`, `n_out`,
-`n_channels` and `rng` can appear literally, with the interpretations explained above. For
-example,
-
-```
-builder = MLJFlux.@builder Chain(Dense(n_in, 128), Dense(128, n_out, tanh))
-```
-
 
 ### Model hyperparameters.
 
@@ -387,7 +378,14 @@ following conditions:
 - The object returned by `chain(x)` must be an `AbstractFloat` vector
   of length `n_out`.
 
-See also `MLJFlux.@builder` for an automated way to create generic builders.
+Alternatively, use `MLJFlux.@builder(neural_net)` to automatically create a builder for
+any valid Flux chain expression `neural_net`, where the symbols `n_in`, `n_out`,
+`n_channels` and `rng` can appear literally, with the interpretations explained above. For
+example,
+
+```
+builder = MLJFlux.@builder Chain(Dense(n_in, 128), Dense(128, n_out, tanh))
+```
 
 ### Loss functions
 
