### 2021-09-27 e3b3033
added new builder for sklearn style mlp
