### 2021-09-27 94244e8
put all layers into hidden and changed sigmoid to relu to match sklearn
