### 2021-09-27 cfa8e6a
updated MLP to put all layers in hidden and changed default hyperparams to match sklearn
