### 2021-09-27 ceeb577
fixed docstring to match new defaults
