### 2021-09-27 11123f4
add disclaimer about training data fitting into memory
