+++ src/builders.jl
@@ -65,7 +65,10 @@ end
 """
     MLP(; hidden=(100,), σ=Flux.relu, rng=GLOBAL_RNG)
 
-MLJFlux builder that constructs a Multi-layer perceptron network. The ith element of `hidden` represents the number of neurons in the ith hidden layer. An activation function `σ` is applied between each layer.
+MLJFlux builder that constructs a Multi-layer perceptron network. The
+ith element of `hidden` represents the number of neurons in the ith
+hidden layer. An activation function `σ` is applied between each
+layer.
 
 The each layer is initialized using `Flux.glorot_uniform(rng)`. If
 `rng` is an integer, it is instead used as the seed for a
@@ -82,7 +85,10 @@ function MLJFlux.build(mlp::MLP, rng, n_in, n_out)
 
     hidden = [Flux.Dense(n_in, mlp.hidden[1], mlp.σ, init=init)]
     for i ∈ 2:length(mlp.hidden)
-        push!(hidden, Flux.Dense(mlp.hidden[i-1], mlp.hidden[i], mlp.σ, init=init))
+        push!(hidden, Flux.Dense(mlp.hidden[i-1],
+                                 mlp.hidden[i],
+                                 mlp.σ,
+                                 init=init))
     end
     push!(hidden, Flux.Dense(mlp.hidden[end], n_out, init=init))
 
