### 2021-09-28 077fe34
re-introduce gpu testing on PR's to dev
